Breakup PoolRow component, tag CTAs
